python-be/requirements.txt
auto-editor
openai-whisper
pydub
google-generativeai
python-dotenv
markdown-it-py
mdit-py-plugins
python-frontmatter
jsonschema
sentence-transformers
numpy
networkx
nltk
